Both primary and secondary specialties are now searched
Before only a physician's primary specialty was considered.
So now, for example, if a DO has a primary specialty in
family medicine and a secondary specialty in sports medicine,
he'll show up in searches for both family medicine and sports
medicine. Previously he would've showed up only in searches
for family medicine. This also goes for the aggregate numbers
in the specialty figure. He'll count as 1 in family medicine
and 1 in sports medicine. Also added new fields to the
physicians database.